DB seed by Tamima

+++ e_com/database/migrations/2021_07_20_171730_create_special_offers_table.php
@@ -16,7 +16,7 @@ class CreateSpecialOffersTable extends Migration
         Schema::create('special_offers', function (Blueprint $table) {
             $table->id();
             $table->string('name');
-            $table->integer('discount');
+            $table->float('discount');
             $table->date('expire_date');
             $table->time('expire_time');
             $table->timestamps();
